+++ src/api/users.ts
@@ -1,9 +1,11 @@
-export interface User {
-  id: number;
-  name: string;
-  username: string;
-  email: string;
-}
+// export interface User {
+//   id: number;
+//   name: string;
+//   username: string;
+//   email: string;
+// }
+
+import { User } from '../types';
 
 const users: User[] = [
   {
