+++ src/types.ts
@@ -1,0 +1,15 @@
+export type TodoWithUser = Todo & { user?: User };
+
+export interface User {
+  id: number;
+  name: string;
+  username: string;
+  email: string;
+}
+
+export interface Todo {
+  id: number;
+  title: string;
+  completed: boolean;
+  userId: number;
+}
